Data.RatioInt

Description

Provides a specialised version of Ratio for Int.

Runs about ten times faster than Ratio while being half as fast as floating-point types.

Documentation

data RatioInt

Rational numbers, with numerator and denominator of the Int type.

(%) :: Int -> Int -> RatioInt infixl 7

Forms the ratio of two Int numbers.